requirements.txt
Flask
SpeechRecognition
gTTS
google-generativeai
